"Camouflage Burster" -> "Stealth Burster"
