--- dat/outfits/utility/camouflage_burster.xml
@@ -1,19 +1,0 @@
-<?xml version='1.0' encoding='UTF-8'?>
-<outfit name="Camouflage Burster">
- <general>
-  <typename>System Improvement</typename>
-  <slot>utility</slot>
-  <size>small</size>
-  <mass>1</mass>
-  <stealth_on />
-  <price>170000</price>
-  <gfx_store>damage_forward.webp</gfx_store>
-  <description>No one really knows the origin of the camouflage burster, however, it seems to have been derived from a more primitive makeshift version used by smugglers to evade patrols. By emitting a barrage of strong signals, it is able to temporarily confuse the sensors of nearby ships and give a significant increase in stealth ability. However, this is only a short temporary effect, and the device takes a long time to recalibrate and load the capacitors.</description>
-  <limit>camouflage burster</limit>
- </general>
- <specific type="modification">
-  <ew_stealth>-45</ew_stealth>
-  <speed_mod>30</speed_mod>
-  <lua>outfits/utility/camouflage_burster.lua</lua>
- </specific>
-</outfit>
