+++ dat/tech/all_outfits.xml
@@ -236,7 +236,7 @@
  <item>Cryogenic Repair Nanobots</item>
  <item>Cloaking Device</item>
  <item>Low Radiation Sensor Modulator</item>
- <item>Camouflage Burster</item>
+ <item>Stealth Burster</item>
  <item>Faraday Tempest Coating</item>
  <item>Lattice Thermal Coating</item>
  <item>Milspec Impacto-Plastic Coating</item>
