+++ dat/tech/criminal_outfits.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <tech name="Criminal Outfits">
- <item>Camouflage Burster</item>
+ <item>Stealth Burster</item>
 </tech>
